requirements.txt
flask
flask-sqlalchemy
ipfsapi
pybase62
pysqlite
